Extracted logic from EspressoDropdownMenu; js_only.
